chore(web): enable end to end ssl (#579)
This enables end to end ssl when using the app/web container. It will
use a self signed developer certificate, which will almost certainly
cause problems when you try and access it via localhost. It should,
however, do the trick just fine when load balancing in AWS.

From now on, though, to access the web ui locally, you need to use
'https://localhost' if you are using the containers.

Signed-off-by: Adam Jacob <adam@systeminit.com>