feat(dal,sdf): create service node/component when application is created (#599)
This change is to (for the moment) support working with a Node in the
frontend without required the node add menu to be fully operational.

A Component of type "service" gets created immediately after
an "application" component is created (i.e. when a user hits the "new
application" link in the UI). It's called "whiskers" to distinguish it
and to be in line with our demo--why not?

Note that we also extracted some common behavior into
`Schema::default_schema_variant_id_for_name`.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>
Co-authored-by: Jacob Helwig <jacob@systeminit.com>

Co-authored-by: Jacob Helwig <jacob@systeminit.com>