feat(web,sdf,dal): node add menu data fetched from schemas (#600)
The node add menu now returns real data calculated from the schemas,
according to the context provided from the frontend. The bulk of the
work is in the `node_menu.rs` module, which is responsible for creating
the `node menu` structure required by the frontend.

It also has an implementation that I'm either proud of or horrified by -
it creates the data structure the frontend needs directly, using a lot
of Arc<RefCell<_>> and one cleverly placed Arc<Mutex<_>>. It does work,
it has test coverage, and I'm pretty sure it can be cleaned up more than
a little bit.

But it's fast enough, and it's time to move on. :)

Special shout out to Fletcher and Jacob, who talked through with me my
options toward the end of the day.

Signed-off-by: Adam Jacob <adam@systeminit.com>