feat(webapp) create script to clean vite's cache (#602)
Sometimes vite fails to clean it's own cache and ignores any changes
made, so we created this tiny script to clean vite's cache to avoid
getting stuck with this issue. Ideally we should debug vite and provide
a PR, but for now we need to focus on other stuff so it's a quick
gambiarra (brazilian hack).