maint(veritech): Match new shape of ResolverFunctionRequest in cyclone (#604)
The shape of ResolverFunctionRequest changed, but we didn't notice due
to this code being behind a feature flag in veritech. The shape now
matches, even though the code is wrong. To get it to not be wrong would
involve a lot more changes than we really need right now to get other
parts of the system working, so fixing the implementation is left as
TODOs.

[sc-1978]