feat(web,sdf,dal): node template from schema in node add (#606)
This integrates the backend Schema information with the hypothetical
node that needs to be created by the Schematic Viewer when a new node is
created.

When a menu item is selected, we then ask the backend for the correct
`Node Template`, which contains all the information you need to render
the node on the schematic, but does not actually create a node (or the
attendant component). The data in the `Node Template` is pulled from the
schema.

There are currently no sockets defined in the schema, nor are we
differentiating between input and output sockets. But we can fix that
when we get to edge integration.

Signed-off-by: Adam Jacob <adam@systeminit.com>
Co-authored-by: Paulo Cabral <paulo@systeminit.com>

Co-authored-by: Paulo Cabral <paulo@systeminit.com>