feat(web,sdf,dal): create nodes from the schematic panel (#609)
Restores creating new nodes from the schematic panel. When the user clicks to drop the new node on the schematic, we trigger a call to create the node (and its attendent component) on the backend.

It also introduces the idea of a `NodeView` on the backend. This object is responsible for taking the data that is persisted to the `Node` in the DAL, along with the data for a `NodeTemplate` in the frontend that comes from the Schema (currently populated as a `NodeTemplate`). This will allow us to create that view for the frontend from any combination of information that makes sense.

Caveats:

* We do not take the position as an argument, since that code was in flight when this was written. Adding it is a good follow-up PR.
* We do not have test coverage for the schematic api endpoints. I assume they will change (perhaps greatly) after Alex updates things. Once they stabilize, we should add tests.
* The actual code to do the create is in the wrong place in the Schematic code. Alex will refactor it.

Co-authored-by: Paulo Cabral <paulo@systeminit.com>
Signed-off-by: Adam Jacob <adam@systeminit.com>

Co-authored-by: Paulo Cabral <paulo@systeminit.com>