bug(dal): Do not unnecessarily cast table name in update_by_id_v1
We were always getting the table name in as text, but internally casting
to a regclass. This was completely unnecessary, and was causing a bug
with trying to use the table name in a where clause
(`WHERE information_schema.columns.table_name = this_table`). Since we
need to use the table name as a string in every place that we are using
it, there's no reason to cast it to a regclass.

Co-authored-by: Nick Gerace <nick@systeminit.com>
Co-authored-by: Adam Jacob <adam@systeminit.com>
Co-authored-by: Fletcher Nichol <fletcher@systeminit.com>