feat(dal,sdf): a `Prop` to system, associated with a `SchemaVariant`
Now `Prop`s get added to `SchemaVariant`s *and* you can edit it in the
frontend schema editor!!

One note: there is an error when interactively adding a new prop in the
schema editor. I bet it's a quick fix, but this is out of scope for this
story right this second, so...sorry?

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>