chore(dal): add edges to the dal
Restores edges to the DAL. Inlines the data for the head and the tail of the object, only supports directed edges (we never had a use case for bidirectionality, so I dropped it), and relies on the schema variants socket information to link together the edge.

Also restores two sockets to every schema variant; an input and an output. This needs to be extended as we restore more functionality.

Signed-off-by: Adam Jacob <adam@systeminit.com>