feat(dal): add functions to dal
This is an initial implementation of functions--known in the codebase as
"Func"--back into the codebase. In prior implementations we had
"resolvers" and this maps roughly to `Func` here.

This change has a set string function that will take a string and return
it, suitable to minimally expose functions into sdf, which is future
work.

References: restore funcs func bindings and fun result [sc-2012]

Signed-off-by: Adam Jacob <adam@systeminit.com>
Co-authored-by: Fletcher Nichol <fletcher@systeminit.com>
Co-authored-by: Jacob Helwig <jacob@systeminit.com>
Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>