chore(dal): Make function binding return value `unprocessed_value` and `value` optional
In order to be able to represent the value for a
`FuncBackendKind::Unset`, we need to be able to store a "nothing" value.
Before this change, we were requiring that `unprocessed_value` and
`value` were a valid JSON object, which would have meant reserving a
value that could have been valid from another backend as our special
"unset" value.

For now, we're loosening the constraint in the DB that
`unprocessed_value` and `value` be present, so that we can use a DB
`null` value to represent "unset". On the Rust side "unset" will be
represented by an Option::None, and all other values become an
Option::Some<T>.

By explicitly storing the "unset" as a `null` in the value, we are still
able to differentiate between "this binding has been executed, and it
explicitly has no value" and "this binding has yet to be executed, so we
don't know what the value is yet".