feat(dal): add AttributeResolver (#639)
This adds AttributeResolver back into the system. It allows you to take
a FuncBinding and say that it should be used to resolve the value of a
given Prop in a particular Context. It features a spiffy function to
look up a FuncBindingResultValue, choosing the correct value based on
the precedence of the various AttributeResolvers contexts that are
relevant for the prop.

It stops short of the full precedence - I suspect we also need to add
systems to the mix. But it's good enough to unblock everyone!

Signed-off-by: Adam Jacob <adam@systeminit.com>