fix(web): fix eslint + most tsc errors
Now eslint only warns about 'any' usage, we should fix this, but first
let's focus on errors and allowing the CI to pass. Then we can make
another expedite task for these warnings.

vue-tsc has a few errors that we can't seem to fix, most related to Vue3
interaction with TS and RxJs, we've managed to reduce from a few hundreds to 23.