feat(dal,web): set value for a string from a function (#645)
This PR connects AttributeResolvers to the EditField implementation for
the Props of Components. It creates an appropriate FuncBinding, and then
ensures that the AttributeResolver is correctly configured for the
context the value is being set in. (In this case, it's directly on the
component).

Lots of future work expanding this - making it work for every prop type,
making it work with function backends, etc. But this gets the job done
for now.

Signed-off-by: Adam Jacob <adam@systeminit.com>
Co-authored-by: Fletcher Nichol <fletcher@systeminit.com>
Co-authored-by: Jacob Helwig <jacob@systeminit.com>