feat(web,dal): return true validations
This PR brings validations fully round trip with the frontend, in a way
that shows how it will work, but is a little janky. :) Essentially, we
create a new ValidationResolver for every property that asserts that the
string should be 'poop', or else have an error. The validation itself is
produced by a Func that returns a ValidationError, proving that we can
set validations for a given context, look them up, and return them to
the frontend.

It's kind of awesome, even if it's a bit of a hack.

Signed-off-by: Adam Jacob <adam@systeminit.com>
Co-authored-by: Jacob Helwig <jacob@systeminit.com>