refactor(cyclone,veritech): unify `OutputStream` & common types
This change combines ResolverFunction and QualificationCheck
`OutputStream` types into a single type and unifies the notion of a
`ProgressMessage` and common result failure types.

The combined types are in `lib/cyclone/src/progress.rs` and most of the
changes are the fallout from there. ;)

References: add the veritech client server call to dispatch [sc-2117]

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>