feat(dal): track function execution, and record any output
Adds a `FuncExecution` module, which is used to track the execution of a
given function. It's analogous to the `HistoryEvent` module, which
tracks user (or system) events for recording their history, except it
does so for function executions.

It will also keep track of the output of a function as it is executing,
and push it down NATS to be consumed by the browser. It does so in a
deeply un-optimized way.

We will eventually revisit how this is factored, I think, as we add more
FuncBackends and put streaming output in the frontend.

Signed-off-by: Adam Jacob <adam@systeminit.com>