fix(data): ensure a NATS subscription can be closed
This change fixes an issue with our Subscription::next_message
implementation. Before this change, the subscription would block (in a
worker thread pool) waiting for the next NATS message, but we forgot to
also expect a shutdown signal which would tell this blocking thread to
abort/quit/end.

The fix was ported from the upstream `nats` crate from the `asynk`
module (which is their async wrapper over their blocking client).

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>