feat(veritech): add graceful shutdown to the veritech server
This changes will gracefully terminate a long run veritech server but
first unsubscribing from all subscriptions before ending the main server
thread. This *also* unlocks the ability to run the veritech server as a
library in unit and integration tests.

In the situation where we might want to programatically shutdown an
instance of the veritech server, there is a `Server::shutdown_tx` method
which returns a shutdown sender.

There are 2 additional changes which are related:

* There were a few Rust types that needed to be re-exported from
  deadpool-cyclone to allow the Veritech server to properly compile on
  its own
* The `#[ignored]` attributes were dropped from the formerly
  non-terminating tests in `lib/veritech/client.rs`

References: Ensure Veritech server can run as a library [sc-2126]

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>