chore(dal,sdf): everyone who wants a Veritech client, gets one!
This change threads through a `veritech::Client` struct to a couple of
DAL model calls that now require it.

As a consequence of having more Veritech clients in our lives moving
forward, this change also backfills a couple of testing macros (namely
`test_setup!` in `dal` and `sdf`) to include a ready-to-go Veritech
client.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>