chore(cyclone,dal,telemetry): make `lang-js` logging configurable
Yes, yes, a task to configure a logging level for `lang-js`...by
changing no code in `lang-js` you say?

Well, `lang-js` is doing just fine, thanks for asking. Right now, a
Cyclone server will spawn a process of `lang-js` and before this change
it was unconditionally setting the output to debug level. This change
threads through a notion of a `TelemetryLevel`, which is a trait in
our `tracing` crate for any type that can answer if it is in debug
level logging or lower (that is, trace, etc.).

Now by default, if a Cyclone server is running a debug or lower logging
level, then any `lang-js` that is spun up will also be at a debug level
of logging.

In our testing code, Cyclone servers and Veritech servers can be
spawned, both of which will ultimately spawn `lang-js` instances.  By
default in tests, these processes will log at their standard info levels
so far less output is seen mixed in with the Rust tests. If you wish to
see `lang-js` output then set the `SI_TEST_VERBOSE` environment variable
to some non-empty value when running tests like `env SI_TEST_VERBOSE=1
cargo test`.

Note that Cyclone and Veritech servers may still be emitting their log
output during tests, but that's the topic of another batch of changes.
Until next time!

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>