bug(web): fix reactivity in attribute panel
Fixes a number of reactivity bugs in the attribute panel. The first was
our SiSelect component not being told to return the values as numbers,
and therefore it emitted values as strings.

The second (and more pernicious) issue was related to when the
observable to fetch the attribute data would fire when it was being fed
property information. It showed itself in two ways:

1. When a component should be reactive to property changes, and that
   reactivity happens from within an observable call (for example, using
   a `service` object to call sdf), the property that you wish to
   observe must be independently turned in to a Ref. (For background,
   the props object itself is reactive, but not *deeply* reactive). This
   can be done with `fromRef(props, { immediate: true, deep: true })`,
   or from otherwise destructuring the props to make them reactive.

2. When trying to use the property value in an observable, there was a
   bug in vuse-rx. The documentation said that calls to `fromRef` could
   take options (as the example above shows). Unfortunately, while the
   types themselves said the function would accept options, the options
   were never passed to the underlying `watch`. The result is that no
   matter what set of options were passed to `fromRef`, the first
   emission would not trigger the watch.

I have patched the upstream library and sent a pull request. When it is
merged and a new release is cut, we can upgrade our dependency. Until
then, we are pointing to a branch with the bugfix in it.

Signed-off-by: Adam Jacob <adam@systeminit.com>