feat(web): Restore Resource sub-panel
This restores a basic ResourceViewer that just dumps the JSON on the screen, without all the specialized sub viewers we used to have. So we can use this to create a generic viewer, that renders any kind of JSON dynamically.

This also creates a mock ResourceService that returns dummy value, to be replaced by an actual SDF call.