feat(veritech): impl resource sync on server & client
This work was easiest to implement together as the verification of the
Veritech server was delivered by a client implementation and test. Also,
I forgot that I hadn't implemented the server before starting on the
client and seeing that my test was failing. Ha.

References: Add resource sync handling to Veritech server [sc-2163]
References: Add resource sync handling to Veritech client [sc-2171]

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>