feat(web,sdf): Implement get/sync resource mock routes
We only want to sync when the user clicks (if not in edit mode).

So we fixed the reactivity and implemented a new route to fetch the last sync result.

Both routes are mocked right now, but it makes the frontend pretty close to the old demo and when we actually implement the routes in the sdf nothing in the frontend ideally should be changed.

There is a catch tho, the way we display the date seems different, we will have to check on how to format the date of the last sync depending on how it comes from the database