feat(dal,sdf): return qualification results via sdf (#701)
Wire qualification checks through to the frontend. This PR links
actually checking the qualification by looking for the
QualificationPrototypes for a given Schema/SchemaVariant, and creating
or updating QualificationResolvers for a given component as appropriate.

It also transforms the FuncBindingResultValue into a QualificationView
suitable for use in the frontend.

It includes some test coverage for qualifications that is fragile - it
depends on the current implementation of the `docker_image` schema. What
it does do is confirm that the full behavior works as expected.

Signed-off-by: Adam Jacob <adam@systeminit.com>