chore(dal,veritech): create default NATS subjects with optional prefixes (#704)
This change adds some more convention around the Veritech server and
client agreeing on the various NATS subjects to use when transmitting
function executions requests. A set of shared functions in the
`veritech` crate are now responsible for building these topics so there
should be very little chance of "mis-wiring" an agreed upon subject over
which to communicate.

This change updates the following:

* Adds a `subject_prefix()` which is optional on the
  `veritech::Config::builder()` (the server's Config). This is used by
  our test suites to inject per-test unique prefixes.
* Adds a `veritech::Client::with_subject_prefix()` alternate constructor
  to configure the client side--this is also used by the test suites.
* Removes the formerly required subject strings on each
  `veritech::Client` method such as `execute_resolver_function()`, etc.
  There are net-new alternate methods such as
  `execute_resolver_function_with_subject()` which take a custom subject
  allowing the caller to override the conventional default.
* The Veritech client has a `subject_prefix()` method, allowing a user
  to determine how the client is set up.
* The Veritech server has a `shutdown_handle()` method which returns a
  type holding a shutdown tx and has a single `shutdown()` method on it.
  The test suite uses this on initial setup to explicitly shut down the
  server used in migrations (more detail below).

Now that we can have concurrently running Veritech with their own
subject namespaces, the dal and sdf test suites can essentially give
each test its own Veritech server instance and pre-configured client
which are wired together. In the `one_time_setup()` function--which is
executed fully only one time at the start of a test suite--a dedicated
Veritech server is spun up to assist with any database and object code
migrations. After the migrations, this server is explicitly asked to
shutdown so the currently running test that ran these migrations doesn't
end up with 2 Veritech servers spawned as tasks in its own tokio
reactor.

This'll be fun...

References: Add default configurable NATS subjects in [sc-2172]

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>