feat(lang-js,cyclone,dal,veritech): implement FuncBackendJsResourceSync
Implement support for the ResourceSyncComponent parameter to ResourceSyncs in lang-js & cyclone & veritech.

Fix lang-js and cyclone ResourceSync tests.

Remove sdf's version of QualificationCheckComponent and use the cyclone's one directly.

Keep ResourceSyncComponent as a different struct than QualificationCheckComponent as we don't know if they will stay the same in the future.

Implement the FuncBackendJsResourceSync and make it executable.

Note: we aren't testing the backend yet, FuncBackends generally are tested in the Resolver/Prototype, we will implement those later today, so this might be slightly broken, but it's an easy fix for when we are actually plugging things together.

We don't actually know what the ResourceSync's return type will be, so it's a dummy for now, and we don't do anything with it yet.

Fix small lint issue I forgot in a previous PR.