fix(dal): in Schematic::find, filter out NodeKind:System nodes
This is likely not the right long term solution, but when logging in to
a new billing account I was getting a error of:

`Error: dal schematic error: system not found`

It's looking like the call of `node.system(...)` is returning `None`
which may mean that the association of node <> system isn't quite 100%.

Also, for this second we don't need system nodes to come back in the
schematic. I presume that we will want all of these nodes eventually, so
this will need to be fixed.

If this solution is too hacky, we can resolve the commented out body
instead (granted I'm looking for a real quick fix so we can use the
interface and am open to alternatives).

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>