feat(sdf,dal): implement sync and integrate Resource(Resolver/Prototype/View)
Create ResourceSyncPrototype, with it generate a Resolver, execute it and with the return value, make a ResourceView and send it to the frontend.

Fix QualificationResolver tests.

Create a builtin ResourceSyncPrototype for docker_schema.

There is a reactivity issue when selecting the resource panel the first time, you have to swap components so the first ResourceView render happens.

The sync happens and the next get will return it's output. But for now we aren't getting a refresh after sync. We need to fix the nats multicast.

Note: We are adding a FuncBindingReturnValue for each execution, which will bloat the db quickly, we need to upsert to it.