feat(dal): Create an "Includes" edge between all created components and the "production" system
Long-term, we'll want to be smarter about picking which `System` we use
for the `Edge`, but for now, we're YOLOing everything into "production".

This also consolidates the "create a `System` with a `Node`, and all the
supporting relations" into `System::new_with_node` similar to how we
have `Component::new_with_schema_variant_with_node`.