feat(dal): qualification metadata (#714)
Adds the qualification metadata. Currently, we use the information
supplied when the Qualification Prototype is created.

Eventually, we want to support changing the metadata from the
qualification function itself, but that is for a separate PR.

We also lay some groundwork for sending the sub-check information that
is required by several of our custom qualification checks in the v0.4
implementation (in particular the way we custom render field validations
and kubeeval checks). (See the `QualificationSubCheck`)

Excelsior!

Signed-off-by: Adam Jacob <adam@systeminit.com>