feat(dal,web): Reactively refresh resource viewer on sync
Backend sends websocket message that resource was synced with the component_id and the system_id.

When the frontend receives it, if the ResourceViewer is opened to that component in that system we fetch the resource.

We use a very bad hack to access the BillingAccountIds from a Component, we need to fix Component belongs_to Workspace in the future.