feat(lang-js): Support executing subprocesses from javascript
This is highly dangerous and should be forbidden when users are creating their customized functions.

To do this we had to make virtually everything about execution async.

We are using NodeVM instead of VM as it supports modules and they seem needed for async retrieval of the return value. NodeVM doesn't support timeouts.

We have to wrap the code in a hackish way to map Promises to VM2 async execution callback (which doesn't support rejections). Albeit we get the advantage of not having to bake the component in the JS code. Using modules means we can pass the argument directly on function call.

For now it waits until the process returns to be able to return anything. We have commented out the old way of streaming the process return value through a WS connection. It's way more complex to support the streaming version and we don't actually need it for now, so we are leaving it there commented to be used in future work.

To test these changes run: `npm run run -- qualificationcheck < examples/qualificationCheckExecTest.json`

This changes kinda hurts usability, because it creates a global siExec object, which when editing the js function, the IDEs don't recognize, so they keep complaining that siExec doesn't exist. And they also complain about the function not being used anywhere. We need to improve this usability in the future, it might be as simple as passing some kind of context to the function as parameter and declaring the function as a module so we don't need to use it.