fix(web): resolve ~48 TypeScript errors and warnings
This change rolls up several TypeScript errors, generated by running
`npm run type-check`.

I'd like to appologize for the massive churn here, but there are
substantially less errors and warnings as a result. Which is not to say
there are zero remaining--we're down to 8 and they somehow are being
sourced from `vuse-rx` which are not type errors if you `npx tsc
--noEmit` check in a clone of `vuse-rx` at the same forked commit.

Problem for another day?

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>