feat(dal): Qualify docker image with skopeo
We install skopeo in the bootstrap script to ensure it's available in the path when needed.

We spawn a subprocess from the qualification check function to skopeo passing the docker image name.

We hardcode the systeminit namespace in docker.

We only support public images, as there is no current way of properly handling secrets

It's not clear if we want to filter the output, for now we pipe stdout on success, stderr on failure.