feat(lang-js/cyclone/veritech/dal): Pass Component Parents to AttributeResolver (#729)
This PR passes the configuration parent components to AttributeResolvers of the child.

Now an attributeResolver set at schema_variant creation will be copied and re-run whenever a prop attached to it is unset. Although there isn't a UI to unset props yet.

In the future we might want to re-run it when a configuration edge is created between two components and when the component is created (as it might not be idempotent).

We are manually building the Component properties on first run (when creating the schema variant), this is not ideal.