chore(ci): Ensure build-image.sh reports Docker image builds
There appears to be something about `docker "${args[@]}"` that is
causing `set -e` to be ineffective. Even when explicitly doing `set -e`
immediately before the `docker` line, the scrip is continuing on, and
not exiting when the `docker` line returns non-zero (at least in my
testing). Adding a `|| die "message"` does get the script to exit and
report the failure as we want, though.