chore(ci): Move docker image builds into single workflow with summary status of all builds
GitHub Actions don't have a good way of saying "only run this workflow
after _all_ of these other workflows have run" (they only have a "run
after _any_ of the workflows in this list have run"). This means that
there isn't a good way of checking that _all_ images have built
successfully while having a 1:1 between workflows and images that are
built.

This changes things so that there is a single "build the images"
workflow, where each image is its own job in the workflow.

The summary job at the end is to enable bors matching against a singular
job name, since it's pattern matching is meant for matching against a
job with a semi-unpredictable name, instead of matching against _all_
jobs with a similar pattern. ([forum post for reference](https://forum.bors.tech/t/bug-wildcard-status-ignores-1-match/438/2))

This maintains the behavior of only building an image if the related
files for it have changed by having a job that first checks which files
have changed, and setting an output that each of the "build image X"
jobs can use to see if they should run. The summary job at the end will
only consider a "build image X" job as a failure if it should have run
in the first place.