bug: fix container builds by disabling auto-build of lang-js
This reverts adding a build.rs script to automatically build lang-js
when cyclone was built. It helped us avoid a footgun in dev, but is
tricky to implement a fix for when it comes time to build the released
docker images. We can have a more permanent fix to both issues (the
footgun and the docker image builds) later.