feat(dal): create a ComponentView that handles nested props
This creates a `ComponentView` that can be created from a `Component`
and optional `System`. It will find all the properties for the schema
variant of the component, along with every value for those properties,
and construct the correct serde_json::Value for the components
properties.

Subsequent PRs will deal with using this construct everywhere, and fix
up a few things related to when values are created for objects and
arrays.

Signed-off-by: Adam Jacob <adam@systeminit.com>