feat(dal): Create Kubernetes Deployment Props
Translate props from `si-registry` to the rust code 1-to-1.

No validation or qualification for now.

We had to disable editing for maps and arrays since they weren't
available, but since we wanted to generate the boilerplate for them in a
schema it was breaking tests.

Since it's too much boilerplate copied by hand it might not be 100%
done, only when generating the YAML and validating it we will be sure.

Note: The Attribute UI for Kubernetes Deployment is completely broken
after this PR (too many fields in a tiny space)