feat(dal): Add default FuncBinding when creating Props
When the Prop is first created, they should have a default value of
being unset. By ensuring that a Prop _always_ has a value, we can have
simpler logic around giving Array/Map/Object Props their appropriate
default value (`[]`/`{}`/`{}`) when their children get a value, and more
specifically, it becomes more consistent that we set their value to
`unset` when none of their children have a value that is not `unset` any
more.