chore(doc): Document Docker version requirement in bootstrap.sh
The currently packaged version of the Docker server (20.10.11) in the
current (21.10), and LTS (20.04) versions of Ubuntu has a rather nasty
networking issue, that can cause a container to lose all external
connectivity, and start responding with "connection refused" until it is
brought down, and back up again. This appears to have been fixed in
Docker 20.10.12, which is available directly from Docker's package
repository.