chore(web): prepare AttributeViewer edit count with static value
This change lands all the work for adding back edit counts except for
the logic to calulate the number.

In the course of landing this work, I discovered a solution to the
`EditFieldComponent`'s repopulating issues. In short, it appears easier
to load/reload/update `EditField`s from the `AttributeViewer` and pass
down the array of `EditField`s as a prop to the `EditFormComponent`
component. You can see the cleanup in that child component and how easy
the logic got which felt like a second symbol that we're possibly on the
right track.

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>