feat(web): Create unset component for widgets
Creates trash icon and unsets widgets when clicked. The backend fails for
now, but the frontend part is done.

There is an open question related to outline when the trashcan is
focused, the demo's default is dotted grey, the current one is blue solid.
I copied the demo as it felt better but it may not be the right approach.

This also fix a reactivity issue when changing components (the widgets
didn't update properly).

It's not clear if Arrays are working properly as unset should work for
inner values, but we can't test it for now.