feat(web): impl EditField edit count in AttributeViewer
As an `EditField` can be a deeply nested data structure, I opted to
solve this with a Visitor pattern over `EditField`s, which lead to a
`ChangedEditFieldCounterVisitor` (maybe too long a name, but it's sort
of clear what's going on).

This approach should make other activities such as calculating the
maximum path depth of an object, as well as any other iterator style
tasks.

References: restore the count of edits in the frontend [sc-2247]

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>