feat(dal): Support Array/Map/Object Props, and tracking their children
For Array and Map properties, we need to keep track of what order the
children of a specific AttributeResolver will appear in, and in the case
of a Map property, what the key is for the value in the child
AttributeResolvers. This is handled by the IndexMap on an
AttributeResolver.

Component.resolve_attribute now also tells us which AttributeResolverId
is associated with the value that it found/created, which helps in
properly resolving the values that should be children of this particular
value, as we'll need to specify which AttributeResolver they are
children of.

When processing the work_queue to build a ComponentView, we now track
the json pointer for each AttributeResolver that we have seen, so that
we can re-use that pointer when we need to insert children of that
AttributeResolver.

Co-authored-by: Adam Jacob <adam@systeminit.com>