feat(dal): Give parent Array/Map/Object a value when giving child a value
If we're setting the child of an Array/Map/Object, we should make sure
the parent Array/Map/Object to has an "empty" value appropriate for its
type (`[]`/`{}`/`{}`).

This does not (appropriately) handle the case where we're unsetting the
value for a particular prop, as we're not yet checking to make sure that
the prop we're unsetting is the last prop to have a value in the set of
children for the parent. This will be handled by a follow-on task.