feat(dal): maps!
Adds working map support to the `ComponentView`. Expands a bit of
surface area so that keys can be passed correctly.

Tests cover only the most basic use of maps. We should write more
comprehensive torture tests once we have a minute to focus on the
ergonomics of creating schemas and setting values.