feat(dal,cyc,ver): Secret selection as a prop
Creates a WidgetKind to allow lazy construction of the Widgets.
Selects need a special kind so we can fill their options.

This PR is a quick-win to implement Widget setup, it's not ideal,
but way less complex than a WidgetPrototype one, which was the
initial intent.

The biggest problem with this PR is that we can't properly set a
default value for a select, as we can't get the select options in the
AttributeResolver.

Fixes a tiny bug in the SelectWidget.

Standard Accessor for Option<Enum(...)> was implemented. Albeit it's not used.