feat(dal): add encrypted secrets
This change introduces the concept of secrets to the system. Currently,
secrets are not too dissimilar from Components in that a `Secret` has an
object type and kind which are currently opaque to the backend. That is,
the message component of the secret is arbitrary data which the front
end will interpret and use as structured data.

A `Secret` currently has one object type of `Credential` with room to
create other types in the future (perhaps single value secret strings,
for example). Additionally, a `Secret` has a kind which helps the API
consumer to understand the message format. An example of one kind is
`DockerHub`, which consists of a `username` and `password` hash.

Secrets are encrypted client side (in our case, the front end in the
browser) and send to the API on creation using the active billing
account's current public key. The API
only ever exposes a `Secret` type which boils down to its name and basic
type metadata, such as `object_type` and `kind`. Notably, neither the
encypted payload nor the unencrypted message will be exposed by the API,
but rather used internally when preparing to invoke functions
in `Cylone`. For this, an `EncryptedSecret` can be fetched from the
database and then decrypted in Rust into a `DecryptedSecret`.

The secrets API allows a encrypted secret to be created, fetched
individually, and as a list. Notably, the API currently does not support
updating a `Secret` (aside from its `name` field) are therefore assumed
to be immutable (until we decide otherwise).

Of special note in the database creation, we have a `VIEW` of `secrets`
which represents a subset of the underlying `encrypted_secrets` table
which should prevent accidental fetching of encrypted payloads into
application memory when making routine "fetch a list of secrets for the
frontend" type calls. The result is that it almost looks like we have an
`EncryptedSecret` DAL object and an additional independant `Secret` DAL
object, which isn't strictly true. Anyway, the resulting Rust module
makes for some interesting reading as the standard model macros and
calls need to be set up very precisely, hence the additional integration
tests.

References: restore the backend part of secrets [sc-2259]

Signed-off-by: Fletcher Nichol <fletcher@systeminit.com>